Use maxvmem instead of vmem for resource limit checks
Ignore-this: 45c1f6c0dc327c44d3c331c210fb27af
This is an attempt make the generation of h_vmem exceeded messages more
reliable in the execd logs. We change from looking at the instantaneous
vmem to the maximum observed through the job. When memory cgroups are
enabled, this allows detection that processes in the job having been
disposed of by the kernel's OOM killer - otherwise the user would just
see an unexplained status 100 / exit_status 137 in the accounting file.

In fact, this might also be important when memory cgroups are not enabled.

However, it does mean that job scripts that attempt to catch and react
to the exit code of processes it launches will no longer be able to do so.

darcs-hash:59bf8d3ca54fede4c71151c1def68f6e06ad7e2f