Fix #1551 Fix do/do not final spool at qmaster shutdown
Ignore-this: 820d4ed8f381787168dadb145463de5
The qmaster is supposed to do a final spool of all objects at shutdown,
unless it is a shutdown because it has detected that a shadow qmaster
has taken over (e.g. by examining act_qmaster).

Unfortunately, the sense of the test was reversed and the final spool was
only done if an active shadow master was detected. This patch restores
the correct sense of the test.

darcs-hash:9b8fba9ac7bce296804a0303810e01621bc5eea3