Fix #435: Correct share tree usage of array jobs Sharetree usage of array jobs was wrong:
Underlying storage keeps per-job usage, not per-task, but
decay_and_sum_usage and delete_debited_job_usage are called
per-task.  Results in:

* While tasks are running, results in usage only acculmulating at the
  rate of the last task in the job array.

* When a task ends, spool forgets the entire job, resulting in the
  entire usage of the last task being added to spool, instead of a
  delta from the last time usage is calculated.

So task array usage climbs too slowly, then jumps at the end of a task -
either to the correct value if no other task is running, or to too
large a value if other tasks are running.

A side-effect is to squashes a few attributes in the spool's
user/project objects that don't need to be there (e.g. I doubt that
the sum of all job submission times is useful).